This patch adds General Freescale CodeWarrior debug support
CONFIG_CW should be defined if the u-boot needs to be debugged.

Signed-off-by: Roy Zang <tie-fei.zang@freescale.com>
Signed-off-by: adrian bogdan <adrian.bogdan@freescale.com>

Add support of kernel debug with attach mode

When uboot downloads and decompresses the kernel.
it jumps at the kernel  entry point.
The debugger needs to have a breakpoint set there,
So it needs the debugger register setting.

Signed-off-by: Roy Zang <tie-fei.zang@freescale.com>
Signed-off-by: adrian bogdan <adrian.bogdan@freescale.com>