powerpc: fix LDSCRIPT in NAND boot
CONFIG_NAND_U_BOOT indicates whether the SPL is used, while
CONFIG_NAND_SPL is set only when building the SPL itself (and isn't
available in makefiles anyway).  This is the linker script for the
main U-Boot when SPL is in use, so CONFIG_NAND_U_BOOT is the
variable to use.

Also, refrain from overriding LDSCRIPT with a NAND-specific script,
unless such a script exists.

Signed-off-by: Zhao Chenhui <b35336@freescale.com>
Acked-by: Scott Wood <scottwood@freescale.com>