powerpc/85xx: remove SERDES4 soft-reset work-around
Some P4080 rev1 errata work-arounds, notably erratum SERDES4, required a
bank soft-reset after the bank was configured and enabled, even though enabling
a bank causes it to reset.  Because the reset was required for multiple errata,
it was not properly enclosed in an #ifdef, and so was not removed with all the
other rev1 errata work-arounds.

Because this was the only SERDES bank soft-reset, there is no need to implement
a work-around for erratum SERDES-A003.

Signed-off-by: Timur Tabi <timur@freescale.com>