16-bit and ECC support for IFC NAND Machine
16-bit buswidth and ECC support for Integrated Flash Controller
NAND Machine, plus some fixes and cleanups.

Note that using hardware ECC on IFC with small-page NAND (which is what
comes on this board) means there will be insufficient OOB space for
JFFS2, since IFC does not support 1-bit ECC.  UBI should work, as it does
not use OOB for anything but ECC.

When hardware ECC is not enabled in CSOR, software ECC is now used.

As any existing environments will be invalid due to the lack of ECC,
take the opportunity to move it up one block and stop wasting a block
due to overspecifying CONFIG_SYS_NAND_U_BOOT_SIZE.

Parts based on code by Dipen Dudhat.

Signed-off-by: Scott Wood <scottwood@freescale.com>