powerpc/fsl-corenet: extend A004510 to other chips, add SVR check
Add workaround to p2041/p3041/p3060/p5020.  Check SVR revision
to avoid applying the workaround on fixed revisions, as the workaround
has a performance penalty.

p2040/p4040/p5010 are not explicitly addressed by this patch as they are
built as p2041/p4080/p5020, respectively -- but the SVR check is designed
to work with them as well.

p5040 was not in the U-Boot tree when this patch was written,
so it is not included here.

Signed-off-by: Scott Wood <scottwood@freescale.com>