powerpc/85xx: add SerDes bank 4 lanes only if we have a bank 4
Only some chips have four SerDes banks, so don't define lanes for a bank
that doesn't exist.

This fixes warning message "excess elements in array initializer" for
array lanes[], because that array is defined with a size of SRDS_MAX_LANES,
and the value of SRDS_MAX_LANES depends on the number of SerDes banks.

Signed-off-by: Timur Tabi <timur@freescale.com>