powerpc/fsl-corenet: fix timebase loop in A-004510 workaround
The timebase was not enabled this early in the boot -- but there
was another bug that prevented the loop from waiting very long.  The actual
delay was 4096 loop iterations (not timebase ticks), until overflow caused
the loop to exit.

The delay is now set to 65536 timebase ticks, which is shorter than
previously intended, but longer than the old accidental delay, which
seemed to be long enough to work.

The total boot delay for this erratum should now be around 20 ms,
assuming a 50 MHz timebase.

Signed-off-by: Scott Wood <scottwood@freescale.com>