e1000e : Set Receive Threshold granularity to number of descriptors in RXDCTL.
In RXDCTL register, WTHRESH controls the write back of processed receive
descriptors. WTHRESH controls the write back of processed receive
descriptors. This threshold refers to the number of receive
descriptors in the Ethernet controllers on-chip buffer which are
ready to be written back to host memory. In the absence of
external events (explicit flushes), the write back occurs only after
more than WTHRESH descriptors are available for write back.
A value of WTHRESH can be in either cache line units, or based on
number of descriptors based on RXDCTL.GRAN. By default it is set to
cache line granularoty.

In e1000 driver, receive descriptor queue is configured such that
hardware can add only 1 descriptor at a time. So the WTHRESH granularity
should be set to single descriptor. This would ensure that every time
controller fills a Rx descriptor, it is flushed to host memory.

This change fixes timeout which was seen when tftping large files.

Signed-off-by: Ruchika Gupta <ruchika.gupta@freescale.com>
Signed-off-by: Vakul Garg <vakul@freescale.com>
Acked-by: Roy Zang <tie-fei.zang@freescale.com>