powerpc/srio: Workaround for srio port AckID mismatch error
SRIO uses Acknowledge Identification (AckID) counters to track each
packet and ensure it has been delivered ultimately to its destination.
Each port keeps track of its AckID counters and these values are used
in the packet acknowledgment control symbol to update the packet
transmit status between devices. AckID is a 5bit counter and, when it
reaches 31, will rollover to 0.

For a SRIO link, the necessary requirement to maintain the correct state
is simply to match the inbound and outbound AckIDs on the both sides
of the link. If the AckIDs do not match, for example to reset SRIO
port when it is transferring the packets with partner, the ports will
be error and stoped.

The core will be hung if it tries to issue SRIO read/write packet by
the AckID mismatch error port.

The workaround will prevent any SRIO read/write access if the port
is stoped due to AckID mismatch error.

Signed-off-by: Liu Gang <Gang.Liu@freescale.com>
Acked-by: Li Yang <leoli@freescale.com>