command/cache: Add flush command
When we copy code/data to the main memory, we may need to flush the
cache if required by architecture. It uses the existing function
flush_cache. Syntax is

flush <addr> <size>

The addr and size are given in hexadecimal. Like memory command, there is
no sanity check for the parameters.

Signed-off-by: York Sun <yorksun@freescale.com>
Change-Id: I8da85db8c72163ac381c0903205ca69985191fd5
Reviewed-on: http://git.am.freescale.net:8181/893
Reviewed-by: Wood Scott-B07421 <scottwood@freescale.com>