powerpc/p5040ds: set pcie liodn in the correct register
The liodn for the new PCIE controller included in P5040DS
is no longer set through a register in the guts register block
but with one in the PCIE register block itself.
Update the PCIE CCSR structure to add the new liodn register
and add a new dedicated SET_PCI_LIODN_BASE macro that puts
the liodn in the correct register.

Signed-off-by: Laurentiu Tudor <Laurentiu.Tudor@freescale.com>