powerpc/p4080ds: Fix PCI-e x8 link training down failure
If we set the PCI-e controller link width as x8 in rcw and add a
x1 PCI-e device to it, it fails to train down to x1 width. This
happens with p4080ds. According to p4080ds errata PCIe-A003, we
reset the PCI-e controller link width to x4 during the PCI-e init
in u-boot. Then it can train down to x1 link width.

Signed-off-by: Yuanquan Chen <B41889@freescale.com>