powerpc/mpc85xx: avoid tlbivax when tlbilx exists
Erratum A-004827 means that tlbivax is unusable on some chips
(e.g. b4420).

Technically the bug is with tlbsync rather than tlbivax, and for some
reason we don't currently do tlbsync after tlbivax when invalidating the
initial locked-cache memory (we don't care about other CPUs here, but
that's also the case in the other invalidation which does use tlbsync),
so we could have gotten away without converting that.  It seemed better
to be consistent, though.

Signed-off-by: Scott Wood <scottwood@freescale.com>