Run A-004857 A-005553 workarounds before L2 is on
The workarounds for errata A-004857 and A-005553 were setting the debug
registers of the first cluster's L2 cache after enabling that cache,
which may result in unexpected behavior.  This patch reimplements the
workarounds to set the first cluster's L2 cache debug registers before
enabling the first cluster's L2 cache.  The workaround for the other
clusters' L2 caches already behaves in this fashion.

Signed-off-by: James Yang <James.Yang@freescale.com>