powerpc/chassis2: Change the way to disable cores
This is a workaround for erratum A006208. When all cores within a cluster
are disabled via COREDISR register, RCPM puts the cores into PH15 then
also triggers the cluster PLL to kill the clock to that cluster while the
cluster is still in PCL0. Without the clock running the cluster stops
responding to CCM transactions thus causes the system to hang. The
workaround is to put cores into PH15 using PCPH15SETR instead of COREDISR
register, or into PH30 using corresponding registers.

Signed-off-by: Zhao Chenhui <chenhui.zhao@freescale.com>
Signed-off-by: York Sun <yorksun@freescale.com>