powerpc/p5040: revert workaround of A-004699
Erratum A-004699:
The SERDES requires the PVCOCNT_EN field of PLL control register 0 to be
set to 1 when running in some configurations where lanes use clocks from
the 'other' bank's PLL (BnGCRm0[UOTHL]=1). The default value of this field
is incorrect in those configurations, and the affected SGMII or XAUI links
in those configurations will not train unless the register value is
corrected.

Impact:
Failure to train SGMII and/or XAUI links if SERDES PLL configuration is
incorrect.

Workaround:
Set SRDSB3PLLCR0[PVCOCNT_EN] = 1'b1
if SRDSB3PLLCR0[POFF]=1 or B3GCRA0[UOTHL]=1 or B3GCRB0[UOTHL]=1
Set SRDSB2PLLCR0[PVCOCNT_EN] = 1'b1
if SRDSB2PLLCR0[POFF]=1 or B2GCRC0[UOTHL]=1 or B2GCRD0[UOTHL]=1

If the above register values are changed after initialization
(i.e. after PBI), the associated bank(s) must be then reset by
setting SRDSBnRSTCTL[RSTREQ]=1.

The workaround was implemented in U-boot without reset serdes bank due to
reset serdes in U-boot is not safe on p5040, but fm2-10g is unstable in this
condition, so we drop the workaround and implement it in PBI.

Signed-off-by: Shaohui Xie <Shaohui.Xie@freescale.com>