fsl_i2c: add workaround for the erratum I2C A00447
This workaround is for the erratum I2C A00447. The device reference manual
provides a scheme that allows the I2C master controller to generate nine
SCL pulses, which enable an I2C slave device that held SDA low to release
SDA. However, due to this erratum, this scheme no longer works. In addition,
when I2C is used as a source of the PBL, the state machine is not able to
recover.

Refer to the erratum A-004447.

Signed-off-by: Zhao Chenhui <chenhui.zhao@freescale.com>