board/bsc9131rdb: Update default boot environment settings
BSC9131RDB has 1GB DDR. Out of this, only 880MB is passed on to Linux via
bootm_size. Remaining
-16MB is reserved for PowerPC-DSP shared control area
-128MB is reserved for DSP private area.

Also 256MB, out of this 880MB is required for data communication between
PowerPC and DSP core. For this bootargs are modified to pass parameter to
create 1 hugetlb page of 256MB via default_hugepagesz, hugepagesz and
hugepages.

Signed-off-by: Priyanka Jain <Priyanka.Jain@freescale.com>
Change-Id: I858324301cd352e5191a67f0feecaf6d1f19542c
Reviewed-on: http://git.am.freescale.net:8181/1431
Reviewed-by: Sun Yusong-R58495 <yorksun@freescale.com>
Tested-by: Sun Yusong-R58495 <yorksun@freescale.com>