powerpc/t4240: add workaround of A005475
A005475:
If a SerDes PLL does not lock when it's expected to lock during the
device reset sequence, a reset request should occur. However, if a
SerDes PLL does not lock, a reset request is not triggered and the
DCFG_RSTRQSR1[SRDS_RST_RR] bit is not set.

Use SRDSxPLLnCR0[PLL_LCK] to determine if the SerDes PLL has locked.
A device reset may be required to obtain SerDes PLL lock.

Signed-off-by: Shaohui Xie <Shaohui.Xie@freescale.com>
Change-Id: Ifa2f8cf111a5e27242f9c818f8e3e371d0731616
Reviewed-on: http://git.am.freescale.net:8181/1312
Reviewed-by: Zang Tiefei-R61911 <tie-fei.zang@freescale.com>
Reviewed-by: Sun Yusong-R58495 <yorksun@freescale.com>
Tested-by: Sun Yusong-R58495 <yorksun@freescale.com>