fsl/usb: Workaround for USB erratum-A005275
Workaround makes FS as default mode on all affected socs.

Add support to check erratum-A005275 validity for an soc. This info is
required to determine whether a given soc is affected by this erratum.
Add quirk for this erratum "has_fsl_erratum_a005275" . This quirk is used
to enable workaround for the errata

Force FS mode as default by:
        - making EPS as FS
        - setting PFSC bit to disable HS chirping

This workaround can be disabled by mentioning "no_erratum_a005275" in
hwconfig string

Signed-off-by: Ramneek Mehresh <ramneek.mehresh@freescale.com>
Change-Id: I6678f26e65e0f4543a4ad3d69002412b961b7180
Reviewed-on: http://git.am.freescale.net:8181/1317
Reviewed-by: Sun Yusong-R58495 <yorksun@freescale.com>
Tested-by: Sun Yusong-R58495 <yorksun@freescale.com>