added version info for ISBC and SFP
T4/B4 SoC's have a different version of ISBC and SFP block.The secure boot
functionality is slightly different with the new version of ISBC and SFP
compared to the earlier SoC's. So the version information is added in file
config_mpc85xx.h
SFP version info which was earlier defined in B4860QDS.h is not a board
specific feature and belongs to SoC header. So the definition is moved to
config_mpc85xx.h

Signed-off-by: Aneesh Bansal <aneesh.bansal@freescale.com>
Change-Id: I09da3e5809f886fadc06821e3439a028e4f2258f
Reviewed-on: http://git.am.freescale.net:8181/1730
Reviewed-by: Wood Scott-B07421 <scottwood@freescale.com>
Reviewed-by: Sun Yusong-R58495 <yorksun@freescale.com>
Tested-by: Sun Yusong-R58495 <yorksun@freescale.com>