powerpc/b4860qds: add workaround for XFI
XFI does not work stable on current board, it's due to heat sink issue,
to make it work stable the board needs additional heat sink, enable two
XFI lanes only, and use Source Optonics optical modules. Right now we
do not have such an erratum for the issue, so use a define
CONFIG_SYS_FSL_B4860QDS_XFI_ERR to identify it. The workaround will only
be used in XFI protocols and only if the hwconfig indicates that XFI is
prefered.

A new VSC3308 config function is used instead of re-use the original
function, to avoid making the function complex and ugly.
It's for SDK1.4, the re-use shall be considered when upstreaming.

Signed-off-by: Shaohui Xie <Shaohui.Xie@freescale.com>
Change-Id: I98e78cf51d834bbc6080a88734d31eaf2e377cb1
Reviewed-on: http://git.am.freescale.net:8181/2825
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Zang Tiefei-R61911 <tie-fei.zang@freescale.com>
Reviewed-by: Sun Yusong-R58495 <yorksun@freescale.com>
Reviewed-on: http://git.am.freescale.net:8181/2937
Reviewed-by: Schmitt Richard-B43082 <B43082@freescale.com>
Tested-by: Schmitt Richard-B43082 <B43082@freescale.com>