powerpc: fix broken cpu "clock-frequency" property
When indexing freqProcessor[] we use the first
value in the cpu's "reg" property, which IDs
threads.
But freqProcessor[] should be indexed with a
core index so, when fixing "the clock-frequency"
cpu node property access the freqProcessor[]
with the core index derived from the "reg' property.
If we don't do this, last half of the "cpu" nodes
will have broken "clock-frequency" values.

Signed-off-by: Laurentiu Tudor <Laurentiu.Tudor@freescale.com>
Change-Id: If9031ee098b6580fa7852fdc814c4ff5c73cfd24
Reviewed-on: http://git.am.freescale.net:8181/5860
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Haiying Wang <Haiying.Wang@freescale.com>
Reviewed-by: Yusong Sun <yorksun@freescale.com>