powerpc:p1010: Adjust memory map to fix SPI/SD boot
1. On-chip boot rom can't load data from ESPI Flash/SD card to L2-Cache

This issue appears due to 3.5G address availability limitation with P1010 boot ROM.
This limitation is there for all SoC which have secure boot ROM even if working in
non-secure mode. Hence the boot loader image to which boot ROM passes on the control
needs to be located with 3.5G address range and also take required actions when it
gets the control with knowledge that it is not located at top of 4G range.

The solution is to change the L2 Cache base from 0xf8f80000 to 0xd0000000.

2. Adjust the layout of L2 Cache for SPI/SD boot, leave enough space for malloc.

Signed-off-by: Zhang Ying <ying.zhang@freescale.com>
Change-Id: Id84288c63ee4f92a52e291251481dbc397346528
Reviewed-on: http://git.am.freescale.net:8181/6600
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Scott Wood <scottwood@freescale.com>
Reviewed-by: Yusong Sun <yorksun@freescale.com>