SECURE BOOT : create tlb entry for CCSR to disable law 0
ISBC creates a LAW 0 entry for non PBL platforms, which is not
disabled before transferring the control to uboot.
The LAW 0 entry has to be disabled.
In the ISBC for P1010 rev 2.0 and 913x rev 1.1 the TLB entry for
CCSR is cleared before exiting the ROM code. So, the TLB entry
of CCSR needs to be created for disabling LAW 0.

Signed-off-by: Aneesh Bansal <aneesh.bansal@freescale.com>
Change-Id: I45f7f0c18edf01b15dba69134565437c7e48f71c
Reviewed-on: http://git.am.freescale.net:8181/6821
Reviewed-by: Ruchika Gupta <ruchika.gupta@freescale.com>
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Yusong Sun <yorksun@freescale.com>