SECURE_BOOT: Disable CPC configured as SRAM earlier in flow
For Secure boot, CPC is configured as SRAM and used as housekeeping
area. This configurations needs to be disabled once in uboot.
Earlier this disabling of CPC as SRAM was happening in cpu_init_r.
As a result cache invalidation function was getting skipped in
case CPC is configured as SRAM. This was resulting in random crashes
on T4 platform.

Signed-off-by: Ruchika Gupta <ruchika.gupta@freescale.com>
Change-Id: Id930ad05edfe98fd71873a9382d33cb95f365114
Reviewed-on: http://git.am.freescale.net:8181/6962
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Yusong Sun <yorksun@freescale.com>