mmc: Update the handling of returned erase block
If the block range was not aligned, we try to align the range size,
The block range actually erased should be less or equal to the block
range sent. If error occured during erase procedure witch part of them
being erased, users should resend the block rang to continue erase the
rest of them.

zero return, has failed.

Signed-off-by: Haijun Zhang <haijun.zhang@freescale.com>
Change-Id: I1153bf59904444a14941e60d2a785022fd7b0818
Reviewed-on: http://git.am.freescale.net:8181/7571
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Xiaobo Xie <X.Xie@freescale.com>
Reviewed-by: Yusong Sun <yorksun@freescale.com>