powerpc/mpc85xx: Add workaround for erratum A007212
Erratum A007212 for DDR is about a runaway condition for DDR PLL
oscilliator. Please refer to erratum document for detail.
For this workaround to work, DDR PLL needs to be disabled in RCW.
However, u-boot needs to know the expected PLL ratio. We put the
ratio in a reserved field RCW[18:23]. U-boot will skip this workaround
if DDR PLL ratio is set, or the reserved field is not set.

Signed-off-by: York Sun <yorksun@freescale.com>
Change-Id: Idea31e4e13ccb12f280de877a450d822a0de8009
Reviewed-on: http://git.am.freescale.net:8181/8447
Reviewed-by: Heinz Wrobel <Heinz.Wrobel@freescale.com>
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>