SECURE BOOT: Add srk key support in esbc
QorIQ platforms like B4/T4 have support of srk key table and
key revocation in ISBC code in Silicon.
The srk key table allows the user to have a key table with multiple
keys andrevoke any key in case of particular key gets compromised.
In case the ISBC code uses the key revocation and srk key table
to verify theu-boot code, the subsequent chain of trust should also
use the same.

The patch adds support for the esbc_validate command to use this
feature in the chain of trust.

Signed-off-by: Gaurav Kumar Rana <gaurav.rana@freescale.com>
Change-Id: I292fe0636e522f7165b8dfff67b1711cc5a61dc3
Reviewed-on: http://git.am.freescale.net:8181/9982
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Aneesh Bansal <aneesh.bansal@freescale.com>
Reviewed-by: Yusong Sun <yorksun@freescale.com>