powerpc/mpc85xx: Remove SoC checking for erratum A007212
Workaround for erratum A007212 applies to selected versions of some
B4/T4 SoCs. It is safe to apply the workaround to all versions. It
is helpful for upgrading SoC without changing u-boot. In case DDR
PLL is disabled by RCW (part of the erratum workaround), we need this
u-boot workround to bring up DDR clock.

Signed-off-by: York Sun <yorksun@freescale.com>
Change-Id: I978f33d6bfa60e2aad960732715fc4ee5f6d3797
Reviewed-on: http://git.am.freescale.net:8181/10087
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>