ARM: HYP/non-sec: Rename CONFIG_SOC_BIG_ENDIAN to CONFIG_PEN_ADDR_BIG_ENDIAN.
Using CONFIG_PEN_ADDR_BIG_ENDIAN will be more exact. For some platform
such LS1, the SoC is LE mode, but some devices and configuration
registers
are in BE mode, so CONFIG_SOC_BIG_ENDIAN is not very exact here.

This update will be found at the V2 of the upstream patch:
  "ARM: HYP/non-sec: add the pen address byte reverting support"

Signed-off-by: Xiubo Li <Li.Xiubo@freescale.com>
Change-Id: I4a15133969cde22bb4b0cd14f12364239e85a921
Reviewed-on: http://git.am.freescale.net:8181/20294
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Chao Fu <B44548@freescale.com>
Reviewed-by: Zhengxiong Jin <Jason.Jin@freescale.com>