ARM: LS1021A: Configure device stream id for smmu.
The chip has 4 SMMU for address translation of the masters. All the
SMMU'shave 8 bit stream ID and SSID. The address translation depends
only on the stream ID of the incoming transaction. The security
attributes and other security features depends on the SSID.
Each master has unique stream ID assigned to it and is configurable
through SCFG registers. The SSID and stream ID for the masters are
identical and share the same register field of STREAM ID registers.

Signed-off-by: Xiubo Li <Li.Xiubo@freescale.com>

---
This patch has sent to the U-boot opensource mail list.
URL: https://patchwork.ozlabs.org/patch/397522/
Change-Id: Id2aeb25cb82312f2dd282f2c8d4df682a95606b3
Reviewed-on: http://git.am.freescale.net:8181/20997
Reviewed-by: Huan Wang <alison.wang@freescale.com>
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Zhengxiong Jin <Jason.Jin@freescale.com>