spi:ls1021aqds:Add dataflash and spi flash compatibility support on LS1021a-qds
Some dataflash (e.g. AT45DB021D) commands is not compatible with normal spi
flashes. This patch will provide a method to judge the spi flash is dataflash
or not. So it will be defined in board file.

The judgement of dataflash:
	int board_spi_is_dataflash(unsigned int bus, unsigned int cs)
The macro of dataflash:
	CONFIG_SF_DATAFLASH

Signed-off-by: Chao Fu <B44548@freescale.com>

The upstream status is pending.
Web link:
http://patchwork.ozlabs.org/patch/393576/

But the future spi framework is under developping, it is in u-boot-dm tree
spi-working branch. The patch will pending until new spi driver framework apllied.
So the compatibility will only adopt with current uboot.

Change-Id: Ice8a0e6703b67ca2b344fa6dd3b37f155d7d67d5
Reviewed-on: http://git.am.freescale.net:8181/22467
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Jingchang Lu <jingchang.lu@freescale.com>
Reviewed-by: Zhengxiong Jin <Jason.Jin@freescale.com>