Secure boot: ISBC Key extension feature added for PBL based platforms
The public key used to verify the signature in the header in
esbc_validate command is embedded in the header itself.A new option
ie_flag (ISBC Key Extension flag) added in the ESBC Headers.
When the flag is set, the public key used to verify signature
is picked up from the ISBC Key Extension Table.

The ISBC Key Extension Table has already been verified by the
internal boot ROM code. The location of the IE Key Table is
the first entry in the SG table in the CSF header of the
verified uboot. The location of the CSF header is
available in the SCFG SCRATCHRW1 register earlier populated by the PBL.

Signed-off-by: Ruchika Gupta <ruchika.gupta@freescale.com>
Change-Id: I3d20cb1267fc1e122276318e28185354d960686d
Reviewed-on: http://git.am.freescale.net:8181/21315
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Yusong Sun <yorksun@freescale.com>