powerpc/b4860qds: add workaround for XFI
XFI does not work stable on current board, it's due to heat sink issue,
to make it work stable the board needs additional heat sink, enable two
XFI lanes only, and use Source Optonics optical modules. Right now we
do not have such an erratum for the issue, so use a define
CONFIG_SYS_FSL_B4860QDS_XFI_ERR to identify it. The workaround will only
be used in XFI protocols and only if the hwconfig indicates that XFI is
prefered.

A new VSC3308 config function is used instead of re-use the original
function, to avoid making the function complex and ugly.

Reviewed-on: http://git.am.freescale.net:8181/2825
Signed-off-by: Shaohui Xie <Shaohui.Xie@freescale.com>
Change-Id: Ic4b1f9c64a82479dcef433c25846ed18b42c2e03
Reviewed-on: http://git.am.freescale.net:8181/23327
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Yusong Sun <yorksun@freescale.com>