powerpc/mpc85xx: use correct dma compatible for several SoCs
Newer qoriq socs have an updated dma ip block with a
different compatible. Let's make sure we use the proper
string so that the dmas get their liodn.
In order to have the means to specify the compatible
string, the liodn setting macros were updated to receive
a new parameter for it.
The following SoCs were changed to use the new compatible:
 T1023/4, T1040, T2080/1, T4240, B4860.

Signed-off-by: Laurentiu Tudor <Laurentiu.Tudor@freescale.com>
Change-Id: Ibb3ec7a13840908685db6e8a27f7e314c3027f38
Reviewed-on: http://git.am.freescale.net:8181/23992
Tested-by: Yusong Sun <yorksun@freescale.com>
Reviewed-by: Yusong Sun <yorksun@freescale.com>