net/phy/realtek: Disable interrupt on Realtek Ethernet PHY drivers
Some Realtek Ethernet PHYs, like RTL8211D(G/N) and RTL8211E(G),
have interrupts enabled by default. If the interrupt is not treated
later by the OS and the PHY's interrupt line is enabled and shared
with other interrupts, the system will get an interrupt storm.
This patch disables the interrupt for PHY devices that use
one of the current Realtek Ethernet PHY drivers.
Some of Realtek Ethernet PHYs, such as RTL8211B(L) have the
interrupt masked. In this case, the functionality of the PHY
should not be afected since this patch brings INER and INSR
registers to their default values.

Signed-off-by: Codrin Ciubotariu <codrin.ciubotariu@freescale.com>
Change-Id: Ic5e353da24f8fec335e446098ed863991a655554
Reviewed-on: http://git.am.freescale.net:8181/30529
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Razvan Stefanescu <razvan.stefanescu@freescale.com>
Reviewed-by: Scott Wood <scottwood@freescale.com>
Reviewed-by: Shaohui Xie <Shaohui.Xie@freescale.com>
Reviewed-by: Prabhakar Kushwaha <prabhakar@freescale.com>