board/t1040d4rdb: Disable all CPLD interrupts, except QSGMI1 and QSGMI2
By default, CPLD used by Freescale's T1040D4RDB has all the interrupt
sources enabled. If the interrupt line is enabled in the OS and one of
these sources rises the interrupt without having a driver to handle it,
then the board will get an interrupt storm.

This patch masks all the interrupts available in CPLD with exception
for QSGMII PHY interrupts, for which we have a driver in Linux.

Signed-off-by: Codrin Ciubotariu <codrin.ciubotariu@freescale.com>
Change-Id: Ie0e78debc2a5c4af2ca2012bfdcd61760eed617f
Reviewed-on: http://git.am.freescale.net:8181/35931
Tested-by: Review Code-CDREVIEW <CDREVIEW@freescale.com>
Reviewed-by: Yusong Sun <yorksun@freescale.com>
Reviewed-by: Priyanka Jain <Priyanka.Jain@freescale.com>
Reviewed-by: Prabhakar Kushwaha <prabhakar@freescale.com>