powerpc/mpc85xx: Add workaround for FMAN erratum A007273
FMAN soft reset is not finished properly if at least one
of the Ethernet MAC clocks is disabled by setting the
corresponding bit of the DCFG_CCSR_DEVDISR2 register.

The workaround is that u-boot with not change DCFG_CCSR_DEVDISR2
where it usually handles the MACS clocks. Then just
before control moves to kernel u-boot will soft reset
fman and only then turn off the clocks of the unused macs.

This erratum applies to the SoCs which are using FMAN V3.

Signed-off-by: Pan Jiafei <Jiafei.Pan@freescale.com>