net: fm: add software reset for erratum A007273
FMAN soft reset is not finished properly if at least one
of the Ethernet MAC clocks is disabled by setting the
corresponding bit of the DCFG_CCSR_DEVDISR2 register.
So in the kernel, FMan soft reset maybe be removed
because of this erratum.
In u-boot, we don't disable any MACs and then add a
FMan soft reset to make sure all version of FMan driver
in kernel works.

Signed-off-by: Pan Jiafei <Jiafei.Pan@nxp.com>